### 2021-09-30 #88 @whedon
[PRE REVIEW]: Designing Machine Learning Pipeline Toolkit for AutoML Surrogate Modeling Optimization
**Submitting author:** @ppalmes (<a href="http://orcid.org/0000-0002-3145-6356">Paulito Palmes</a>)
**Repository:** <a href="https://github.com/IBM/AutoMLPipeline.jl" target ="_blank">https://github.com/IBM/AutoMLPipeline.jl</a>
**Version:**
