+++ paper/paper.tex
@@ -32,7 +32,7 @@
 \usepackage{hyperref}
 \newcommand{\theHalgorithm}{\arabic{algorithm}}
 
-
+ 
 \begin{document}
 
 \input{header}
