+++ paper/twoblocks-all-one.tex
@@ -1,6 +1,5 @@
 \begin{table*}[htbp]
   \tbl{\emph{All-one} (two-block)}{
-  \label{tab:twoblocks-all-one}
   \centering
   \tiny
   \begin{tabular}{{@{} rcrrr|ll|ll|l @{}}}
@@ -25,4 +24,5 @@
 	 Median/Mode  & 1.0  & 25.37  & 2.70 & 2.62  & stdsc    & noop  & noop     & noop  & rf\\
 	 \bottomrule
   \end{tabular}}
+  \label{tab:twoblocks-all-one}
 \end{table*}
