+++ paper/oneblock-all-all.tex
@@ -1,6 +1,5 @@
 \begin{table*}[htbp]
   \tbl{\emph{All-all} (one-block) search strategy}{
-  \label{tab:singleblock}
   \centering
   \tiny
   \begin{tabular}{{@{} rcrrrlll @{}}}
@@ -25,4 +24,5 @@
 	 Median           & 1.5  & 27.92  & 2.54 & 0.23 &            &      & \\
     \bottomrule
   \end{tabular}}
+   \label{tab:singleblock}
 \end{table*}
