+++ paper/tmp/author.tex
@@ -1,27 +1,0 @@
-\author{%
-  Paulito P.~Palmes\\
-  IBM Research \\
-  Mulhuddart, Dublin 15, Ireland \\
-  \texttt{paulpalmes@ie.ibm.com} \\
-  \And
-  Akihiro Kishimoto\\
-  IBM Research \\
-  Mulhuddart, Dublin 15, Ireland \\
-  \texttt{akihirok@ie.ibm.com}
-  \And
-  Radu Marinescu\\
-  IBM Research \\
-  Mulhuddart, Dublin 15, Ireland \\
-  \texttt{radu.marinescu@ie.ibm.com}\\
-  \And 
-  Parikshit Ram\\
-  IBM Research\\
-  Sandy Springs, GA, United States\\
-  \texttt{parikshit.ram@ibm.com}\\
-  \And
-  Elizabeth Daly\\
-  IBM Research \\
-  Mulhuddart, Dublin 15, Ireland \\
-  \texttt{elizabeth.daly@ie.ibm.com}
-}
-
