+++ paper/complexity-surrogate.tex
@@ -1,6 +1,5 @@
 \begin{table*}[htbp]
   \tbl{\emph{All-one with Surrogates}}{
-  \label{tab:surrogate}
   \centering
   \tiny
   \begin{tabular}{{@{} r|crrr|crrr|crrr| @{}}}
@@ -24,4 +23,5 @@
 	 Median       & 1.0  & 26.05  & 2.77 & 2.28   & 1.0  & 31.58  & 3.55 & 0.07   &  2.0 & 35.63  & 3.68 & 0.04 \\
 	 \bottomrule
   \end{tabular}}
+  \label{tab:surrogate}
 \end{table*}
