+++ paper/oneblock-skrf.tex
@@ -1,7 +1,6 @@
 \begin{table*}[htbp]
   \centering
   \tbl{\emph{One-all} (one-block)}{
-  \label{tab:oneblock-skrf}
   \tiny
   \begin{tabular}{{@{} rcrrrlll @{}}}
 	 \toprule
@@ -25,4 +24,5 @@
 	 Median           & 2.50 & 29.84  & 3.69 & 0.11 &          &       & \\
 	 \bottomrule
   \end{tabular}}
+   \label{tab:oneblock-skrf}
 \end{table*}
