+++ paper/openmldata.tex
@@ -2,7 +2,6 @@
   \centering
   \tiny
   \tbl{OpenML Datasets}{
-  \label{tab:openmldata}
   \begin{tabular}{{@{} rccccccccc @{}}}
     \toprule
 	 Dataset      & NAs  & Size & Cat & Num & Rows  & Cols & Class & Best Result       & AutoML \\
@@ -21,4 +20,5 @@
 	 soybean      & 2337 & 171K & 36  & 0   & 683   & 36   & 19    & 06.59 $\pm$ 02.79 & ATM \\
     \bottomrule
   \end{tabular}}
+   \label{tab:openmldata}
 \end{table*}
