+++ paper/oneblock-all-one.tex
@@ -1,7 +1,6 @@
 \begin{table*}[htbp]
   \centering
   \tbl{\emph{All-one} (one-block)}{
-  \label{tab:all-one-single}
   \tiny
   \begin{tabular}{{@{} rcrrrlll @{}}}
 	 \toprule
@@ -25,4 +24,5 @@
 	 Median           & 1.5  & 27.75  & 3.57 & 0.05 &          &      & \\
 	 \bottomrule
   \end{tabular}}
+  \label{tab:all-one-single}
 \end{table*}
