+++ MAINTAINERS.md
@@ -1,13 +1,0 @@
-# Main author
-- [Paulito P. Palmes, IBM Research](https://github.com/ppalmes)
-
-# Contributors
-This is an open-source project sponsored by IBM. We highly encourage contribution by making a PR.
-
-# Pull Requests
-We recommend contributors to submit a PR directly in Github to have public discussion of issues
-for transparency. When providing feedback, please practice a good sense of civility, avoid personal
-attacks and abusive comments.
-
-# Merge Approval
-Make sure the PR has good documentation of the why/what issues are being addressed or features added.
