+++ Project.toml
@@ -1,29 +1,0 @@
-name = "AutoMLPipeline"
-uuid = "08437348-eef5-4817-bc1b-d4e9459680d6"
-authors = ["Paulito Palmes <ppalmes@gmail.com>"]
-version = "0.3.6"
-
-[deps]
-AMLPipelineBase = "e3c3008a-8869-4d53-9f34-c96f99c8a2b6"
-Conda = "8f4d0f93-b110-5947-807f-2305c1781a2d"
-DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
-PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
-Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[compat]
-AMLPipelineBase = "0.1"
-Conda = "1.0, 1.1, 1.2, 1.3, 1.4, 1.5"
-DataFrames = "0.17, 0.18, 0.19, 0.20, 0.21, 0.22, 1.0, 1.1"
-PyCall = "1.90, 1.91, 1.92"
-julia = "1"
-
-[extras]
-Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
-PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
-Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[targets]
-test = ["Dates", "Test", "Random", "PyCall", "Statistics"]
