+++ travis.yml
@@ -1,46 +1,0 @@
-# Documentation: http://docs.travis-ci.com/user/languages/julia/
-language: julia
-os:
-  - linux
-  - osx
-  - windows
-julia:
-  - 1.6
-  - nightly
-addons:
-  apt:
-    packages:
-      - python3-numpy
-      - python3-scipy
-env:
-  - PYTHON=Conda LD_LIBRARY_PATH=$LD_LIBRARY_PATH:$HOME/.julia/conda/3/lib
-
-branches:
-   only:
-      - master
-      - main
-      - stable
-   except:
-      - experimental
-
-notifications:
-  email: false
-
-after_success:
-  - if [[ $TRAVIS_OS_NAME = linux ]]; then
-      julia --project=coverage/ -e 'using Pkg; Pkg.instantiate();
-          using Coverage; Codecov.submit(Codecov.process_folder())';
-    fi
-
-jobs:
-  allow_failures:
-    - julia: nightly
-  include:
-    - stage: "Documentation"
-      julia: 1.5
-      os: linux
-      script:
-        - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
-        - julia --project=docs/ docs/make.jl
-      name: "HTML"
-      after_success: skip
