### 2021-10-07 #88 @matbesancon
seems the error is linked to a missing package now
