+++ paper/paper.tex
@@ -7,7 +7,7 @@
 \usepackage{amssymb}
 
 \usepackage{amsfonts}       % blackboard math symbols
-\usepackage{nicefrac}       % compact symbols for 1/2, etc.
+%\usepackage{nicefrac}       % compact symbols for 1/2, etc.
 \usepackage{microtype}      % microtypography
 %\usepackage[usecolors=true]{jlcode}    % julia code
 \usepackage{microtype}
