+++ paper/paper.tex
@@ -8,9 +8,8 @@
 
 \usepackage{amsfonts}       % blackboard math symbols
 %\usepackage{nicefrac}       % compact symbols for 1/2, etc.
-\usepackage{microtype}      % microtypography
+%\usepackage{microtype}      % microtypography
 %\usepackage[usecolors=true]{jlcode}    % julia code
-\usepackage{microtype}
 \usepackage{graphicx}
 \usepackage{subfigure}
 \usepackage{url}            % simple URL typesetting
