+++ paper/paper.tex
@@ -10,8 +10,8 @@
 %\usepackage{nicefrac}       % compact symbols for 1/2, etc.
 %\usepackage{microtype}      % microtypography
 %\usepackage[usecolors=true]{jlcode}    % julia code
+%\usepackage{subfigure}
 \usepackage{graphicx}
-\usepackage{subfigure}
 \usepackage{url}            % simple URL typesetting
 \usepackage{booktabs}
 \usepackage{tabularx}
