+++ paper/paper.tex
@@ -11,7 +11,7 @@
 %\usepackage{microtype}      % microtypography
 %\usepackage[usecolors=true]{jlcode}    % julia code
 \usepackage{graphicx}
-\usepackage{subfigure}
+%\usepackage{subfigure}
 \usepackage{url}            % simple URL typesetting
 \usepackage{booktabs}
 \usepackage{tabularx}
