### 2021-10-07 #88 @vchuravy
@ppalmes you can use https://whedon.theoj.org/ to get access to the logs.

https://whedon.theoj.org/preview?id=1f1442aa9bb8bf86c6350bbb
```
Your paper failed to compile with the following errors:

Looks like we failed to compile the PDF with the following error: [WARNING] Could not load include file 'amlpcode' at line 225 column 19 [WARNING] Could not load include file 'amlpcode' at line 225 column 19 [WARNING] Could not load include file 'amlpcode' at line 225 column 19 [WARNING] Could not load include file 'amlpcode' at line 225 column 19 Error producing PDF. ! LaTeX Error: Can be used only in preamble. See the LaTeX manual or LaTeX Companion for explanation. Type H for immediate help. ... l.313 \documentclass
```
