+++ paper/paper.tex
@@ -256,7 +256,7 @@ Weekday (C), and Overtime (C).
 \vskip 6pt
 
 \begin{lstfloat}[htbp]
-  \lstinputlisting{amlpcode}
+  \lstinputlisting{amlpcode.tex}
   \caption{AMLP toolkit sample usage}
   \label{lst:amlpcode}
 \end{lstfloat}
