### 2021-10-08 f89964c
minor edits of presentation links
