### 2021-11-04 b8a5849
kubernetes example deployment of automlpipline
