### 2021-11-04 7c6cd21
kubernetes example deployment of automlpipline
