+++ examples/kubernetes/automlpipeline.yaml
@@ -1,0 +1,13 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  generateName: automlpipeline-
+spec:
+  serviceAccountName: "julia-manager-serviceaccount"
+  restartPolicy: Never
+  containers:
+  - name: manager
+    image: "ppalmes/amlp:latest"
+    imagePullPolicy: Always
+    command: ["bash","-c","export JULIA_WORKER_TIMEOUT=500; julia --project main.jl 10 1"]
+    #args: ["${ARG}"]
