+++ examples/juliacon2021/main.jl
@@ -7,14 +7,10 @@ nprocs() == 1 && addprocs(; exeflags = "--project")
 @everywhere include("twoblocks.jl")
 @everywhere using Main.TwoBlocksPipeline
 
-#dataset = getiris()
-#X = dataset[:,1:4]
-#Y = dataset[:,5] |> collect
-#X = dataset[:,2:end-1]
-
-dataset=CSV.read("environment_data.csv",DataFrame)
-X = select(dataset, Not([:state,:time]))
-Y = dataset.state
+dataset = getiris()
+X = dataset[:,1:4]
+Y = dataset[:,5] |> collect
+X = dataset[:,2:end-1]
 
 results=TwoBlocksPipeline.twoblockspipelinesearch(X,Y)
 println("best pipeline is ", results[1]," with mean-sd ",results[2], " ± ",results[3])
