### 2021-11-04 #88 @ablaom
I have now read the manuscript but have not attempted to use the
software.

### Synopsis

A Julia software package called AMLP is described. This software
allows one to combine multiple pre-processing elements (categorical
encoders, dimension reducers, etc) with a supervised learner to
construct a new supervised learner pipeline, with a view to searching for
combinations that maximize a performance estimate. The individual
elements are models wrapped from scikit-learn or R. This architecture
searching tool is similar to existing tools (such as TPOT and AutoSk),
in the sense of what kind of architectures are possible. However, as a
Julia implementation, it appears to have advantages in terms of
usability and performance.

I think this work and software is a valuable contribution and would be
of interest to the Julia community. I do have some concerns I
would like addressed before recommending publication:

#### Methodology for comparing frameworks

In section 3.2 (see especially Tables 2-7) AMLP is ranked, compared
with its competitors, on various datasets, according to the average
performance estimate `AvgErr`. As I understand it, no account is taken
of the *uncertainties* of these estimates, which means the rankings
are, by themselves not very useful (at least, as they apply to a
single dataset).

For example it may be that framework A is ranked ahead of framework B
on a given data set because A gets an `AvgErr` of 76, while B
scores 78. But if estimates for the associated standard errors are +/-
5.0, the difference is not statistically significant, and we should
rank both A and B the same (i.e., call a draw).

Perhaps some hypothesis test based face-off or similar would be more
meaningful? See for example [this
paper](https://link.springer.com/article/10.1007%2Fs10994-006-6226-1).

Perhaps I am missing something? Can the authors clarify this point?

#### Characterization of MLJ

The following characterization of the MLJ toolbox (Under "1.1 Related
Work") has never been true:

"However, [MLJ's] pipeline model only supports a linear non-branching chain of transformers and learners"

Since version 0.1, MLJ has allowed the user to build composite models
of considerably greater complexity than the pipelines constructible
using AMLP. To give two examples, one can construct, using MLJ's
generic composite model syntax, a Wolpert model stack (see
[here](https://github.com/ablaom/MachineLearningInJulia2020/blob/master/wow.ipynb))
and one can wrap supervised models with *learned* target
transformations (see
[here](https://alan-turing-institute.github.io/MLJ.jl/dev/composing_models/#Building-a-simple-learning-network)). Such
model composition features are explicated in [the
documentation](https://alan-turing-institute.github.io/MLJ.jl/dev/about_mlj/#Model-composability)
and in these two articles: [general
paper](https://doi.org/10.21105/joss.02704) and [model composition
paper](https://arxiv.org/abs/2012.15505).

Additionally, I find the following statement misleading:

"Also, [MLJ's] pipeline architecture does not support symbolic
expression for easy to understand pipeline composition similar to
AMLP."

It seems to me that MLJ's canned [`@pipeline`
macro](https://alan-turing-institute.github.io/MLJ.jl/dev/linear_pipelines/)
provides a way to construct some kinds of pipelines using a "symbolic
expression" that is "easy to understand". There isn't the `+` operator
for recombining "select-and-transform" operations (a nice feature) but
you can sometimes achieve the same results by modifying
hyperparameters of models that offer this functionality internally
(e.g., `ContinuousEncoder`).
