+++ test/runtests.jl
@@ -8,6 +8,7 @@ warnings = pyimport("warnings")
 warnings.filterwarnings("ignore")
 
 # test modules
+include("test_basefilter.jl")
 include("test_skpreprocessing.jl")
 include("test_sklearner.jl")
 include("test_skcrossvalidator.jl")
