+++ .github/workflows/ci.yml
@@ -58,14 +58,14 @@ jobs:
             ${{ runner.os }}-test-
             ${{ runner.os }}-
 
-        #      - uses: julia-actions/julia-buildpkg@v1
-        #        env: 
-        #          PYTHON: ""
-        #      - uses: julia-actions/julia-runtest@v1
-        #      - uses: julia-actions/julia-processcoverage@v1
-        #      - uses: codecov/codecov-action@v1
-        #        with:
-        #          file: lcov.info
+      - uses: julia-actions/julia-buildpkg@v1
+        env: 
+          PYTHON: ""
+      - uses: julia-actions/julia-runtest@v1
+      - uses: julia-actions/julia-processcoverage@v1
+      - uses: codecov/codecov-action@v1
+        with:
+          file: lcov.info
   docs:
     name: Documentation
     runs-on: ubuntu-latest
