+++ README.md
@@ -2,7 +2,7 @@
 
 | **Documentation** | **Build Status** | **Help** |
 |:---:|:---:|:---:|
-| [![][docs-dev-img]][docs-dev-url] [![][docs-stable-img]][docs-stable-url] | [![][travis-img]][travis-url] | [![][slack-img]][slack-url] [![][gitter-img]][gitter-url] |
+| [![][docs-dev-img]][docs-dev-url] [![][docs-stable-img]][docs-stable-url] | [![][travis-img]][travis-url] |[![][travis-img]][travis-url] [![][codecov-img]][codecov-url]  | [![][slack-img]][slack-url] [![][gitter-img]][gitter-url] |
 
 ### Stargazers over time
 
