### 2021-12-09 ed9ff85
fix minor bug in demo
