+++ test/runtests.jl
@@ -11,5 +11,6 @@ warnings.filterwarnings("ignore")
 include("test_skpreprocessing.jl")
 include("test_sklearner.jl")
 include("test_skcrossvalidator.jl")
+include("test_basefilter.jl")
 
 end
