+++ test/runtests.jl
@@ -8,8 +8,10 @@ warnings = pyimport("warnings")
 warnings.filterwarnings("ignore")
 
 # test modules
+
 include("test_skpreprocessing.jl")
 include("test_sklearner.jl")
 include("test_skcrossvalidator.jl")
 
+include("test_basefilter.jl")
 end
