+++ test/runtests.jl
@@ -12,4 +12,7 @@ include("test_skpreprocessing.jl")
 include("test_sklearner.jl")
 include("test_skcrossvalidator.jl")
 
+
+include("test_basefilter.jl")
+
 end
