+++ test/runtests.jl
@@ -7,11 +7,8 @@ using PyCall
 warnings = pyimport("warnings")
 warnings.filterwarnings("ignore")
 
-# test modules
-
 include("test_skpreprocessing.jl")
 include("test_sklearner.jl")
 include("test_skcrossvalidator.jl")
-
 include("test_basefilter.jl")
 end
