+++ CondaPkg.toml
@@ -1,0 +1,6 @@
+channels = ["mkl", "scikit-learn"]
+
+[deps]
+scikit-learn = ""
+nomkl = ""
+python = ""
