### 2022-06-24 25ee13f
adding support of PythonCall in crossvalidation
