### 2022-06-24 f401ca0
adding support of PythonCall in crossvalidation
