+++ Project.toml
@@ -16,15 +16,15 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 AMLPipelineBase = "0.1"
 Conda = "1.0, 1.1, 1.2, 1.3, 1.4, 1.5"
 DataFrames = "0.17, 0.18, 0.19, 0.20, 0.21, 0.22, 1.0, 1.1"
-PyCall = "1.90, 1.91, 1.92"
+PythonCall = "0.9"
 julia = "1"
 
 [extras]
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
-PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
+PythonCall = "6099a3de-0909-46bc-b1f4-468b9a2dfc0d"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Dates", "Test", "Random", "PyCall", "Statistics"]
+test = ["Dates", "Test", "Random", "PythonCall", "Statistics"]
