+++ Project.toml
@@ -15,6 +15,7 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 [compat]
 AMLPipelineBase = "0.1"
 Conda = "1.0, 1.1, 1.2, 1.3, 1.4, 1.5"
+CondaPkg = "0.2"
 DataFrames = "0.17, 0.18, 0.19, 0.20, 0.21, 0.22, 1.0, 1.1"
 PythonCall = "0.9"
 julia = "1"
