### 2022-06-24 49f8831
fix crossvalidation tests
