### 2022-06-24 297f19c
bug fixes to support README examples
