### 2022-06-24 71b94ca
bug fixes to support README examples
