+++ Project.toml
@@ -5,7 +5,6 @@ version = "0.4.0"
 
 [deps]
 AMLPipelineBase = "e3c3008a-8869-4d53-9f34-c96f99c8a2b6"
-Conda = "8f4d0f93-b110-5947-807f-2305c1781a2d"
 CondaPkg = "992eb4ea-22a4-4c89-a5bb-47a3300528ab"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 PythonCall = "6099a3de-0909-46bc-b1f4-468b9a2dfc0d"
@@ -14,7 +13,6 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [compat]
 AMLPipelineBase = "0.1"
-Conda = "1.0, 1.1, 1.2, 1.3, 1.4, 1.5"
 CondaPkg = "0.2"
 DataFrames = "0.17, 0.18, 0.19, 0.20, 0.21, 0.22, 1.0, 1.1"
 PythonCall = "0.9"
