+++ dev/assets/documenter.js
@@ -72,6 +72,67 @@ $(document).ready(function() {
 
 })
 ////////////////////////////////////////////////////////////////////////////////
+require([], function() {
+function addCopyButtonCallbacks() {
+  for (const el of document.getElementsByTagName("pre")) {
+    const button = document.createElement("button");
+    button.classList.add("copy-button", "fas", "fa-copy");
+    el.appendChild(button);
+
+    const success = function () {
+      button.classList.add("success", "fa-check");
+      button.classList.remove("fa-copy");
+    };
+
+    const failure = function () {
+      button.classList.add("error", "fa-times");
+      button.classList.remove("fa-copy");
+    };
+
+    button.addEventListener("click", function () {
+      copyToClipboard(el.innerText).then(success, failure);
+
+      setTimeout(function () {
+        button.classList.add("fa-copy");
+        button.classList.remove("success", "fa-check", "fa-times");
+      }, 5000);
+    });
+  }
+}
+
+function copyToClipboard(text) {
+  // clipboard API is only available in secure contexts
+  if (window.navigator && window.navigator.clipboard) {
+    return window.navigator.clipboard.writeText(text);
+  } else {
+    return new Promise(function (resolve, reject) {
+      try {
+        const el = document.createElement("textarea");
+        el.textContent = text;
+        el.style.position = "fixed";
+        el.style.opacity = 0;
+        document.body.appendChild(el);
+        el.select();
+        document.execCommand("copy");
+
+        resolve();
+      } catch (err) {
+        reject(err);
+      } finally {
+        document.body.removeChild(el);
+      }
+    });
+  }
+}
+
+if (document.readyState === "loading") {
+  document.addEventListener("DOMContentLoaded", addCopyButtonCallbacks);
+} else {
+  addCopyButtonCallbacks();
+}
+
+})
+////////////////////////////////////////////////////////////////////////////////
 require(['jquery', 'headroom', 'headroom-jquery'], function($, Headroom) {
 
 // Manages the top navigation bar (hides it when the user starts scrolling down on the
@@ -222,6 +283,12 @@ require(['jquery'], function($) {
 
 // update the version selector with info from the siteinfo.js and ../versions.js files
 $(document).ready(function() {
+  // If the version selector is disabled with DOCUMENTER_VERSION_SELECTOR_DISABLED in the
+  // siteinfo.js file, we just return immediately and not display the version selector.
+  if (typeof DOCUMENTER_VERSION_SELECTOR_DISABLED === 'boolean' && DOCUMENTER_VERSION_SELECTOR_DISABLED) {
+    return;
+  }
+
   var version_selector = $("#documenter .docs-version-selector");
   var version_selector_select = $("#documenter .docs-version-selector select");
 
