### 2022-06-29 312b642
set folds to 3 for faster demo
