+++ examples/k8s-arm/twoblocks.jl
@@ -1,0 +1,128 @@
+module TwoBlocksPipeline
+
+export twoblockspipelinesearch
+
+using Distributed
+#nprocs() == 1 && addprocs(; exeflags = "--project")
+#nprocs() == 1 && addprocs()
+@everywhere using AutoMLPipeline
+@everywhere using DataFrames
+@everywhere using DataFrames:DataFrame
+@everywhere using AutoMLPipeline: score
+@everywhere using Random
+
+# disable truncation of dataframes columns
+import Base.show
+show(df::AbstractDataFrame) = show(df,truncate=0)
+show(io::IO,df::AbstractDataFrame) = show(io,df;truncate=0)
+
+# define scalers
+const rb     = SKPreprocessor("RobustScaler",Dict(:name=>"rb"))
+const pt     = SKPreprocessor("PowerTransformer",Dict(:name=>"pt"))
+const norm   = SKPreprocessor("Normalizer",Dict(:name=>"norm"))
+const mx     = SKPreprocessor("MinMaxScaler",Dict(:name=>"mx"))
+const std    = SKPreprocessor("StandardScaler",Dict(:name=>"std"))
+# define extractors
+const pca    = SKPreprocessor("PCA",Dict(:name=>"pca"))
+const fa     = SKPreprocessor("FactorAnalysis",Dict(:name=>"fa"))
+const ica    = SKPreprocessor("FastICA",Dict(:name=>"ica"))
+# define learners
+const rf     = SKLearner("RandomForestClassifier",Dict(:name => "rf"))
+const ada    = SKLearner("AdaBoostClassifier",Dict(:name => "ada"))
+const gb     = SKLearner("GradientBoostingClassifier",Dict(:name => "gb"))
+const lsvc   = SKLearner("LinearSVC",Dict(:name => "lsvc"))
+const rbfsvc = SKLearner("SVC",Dict(:name => "rbfsvc"))
+const dt     = SKLearner("DecisionTreeClassifier",Dict(:name =>"dt"))
+# preprocessing
+const noop = Identity(Dict(:name =>"noop"))
+const ohe  = OneHotEncoder(Dict(:name=>"ohe"))
+const catf = CatFeatureSelector(Dict(:name=>"catf"))
+const numf = NumFeatureSelector(Dict(:name=>"numf"))
+
+const vscalers =    [rb,pt,norm,mx,std,noop]
+const vextractors = [pca,fa,ica,noop]
+const vlearners =   [rf,gb,lsvc,rbfsvc,ada,dt]
+const learnerdict = Dict("rf"=>rf,"gb"=>gb,"lsvc"=>lsvc,"rbfsvc"=>rbfsvc,"ada"=>ada,"dt"=>dt)
+
+
+function oneblock_pipeline_factory(scalers,extractors,learners)
+   results = @distributed (vcat) for lr in learners
+      @distributed (vcat) for xt in extractors
+         @distributed (vcat) for sc in scalers
+            # baseline preprocessing
+            prep = @pipeline ((catf |> ohe) + numf)
+            # one-block prp
+            expx = @pipeline prep |> (sc |> xt) |> lr
+            scn   = sc.name[1:end - 4];xtn = xt.name[1:end - 4]; lrn = lr.name[1:end - 4]
+            pname = "($scn |> $xtn) |> $lrn"
+            DataFrame(Description=pname,Pipeline=expx)
+         end
+      end
+   end
+   return results
+end
+
+function evaluate_pipeline(dfpipelines,X,Y;folds=3)
+   res=@distributed (vcat) for prow in eachrow(dfpipelines)
[…]
