+++ examples/k8s-arm/Project.toml
@@ -1,0 +1,9 @@
+[deps]
+AbstractTrees = "1520ce14-60c1-5f80-bbc7-55ef81b5835c"
+ArgParse = "c7e460c6-2fb9-53a9-8c5b-16f535851c63"
+AutoMLPipeline = "08437348-eef5-4817-bc1b-d4e9459680d6"
+CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
+DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
+Distributed = "8ba89e20-285c-5b6f-9357-94700520ee1b"
+K8sClusterManagers = "5aeab163-63d2-4171-9fbf-e22244d80acb"
+Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
