### 2022-07-03 7f9feb8
add more info about k8s cluster setup
