### 2022-07-03 a3e22ba
add more info about k8s cluster setup
