+++ examples/juliacon2021/main.jl
@@ -1,7 +1,9 @@
 using Distributed
 using DataFrames
 using CSV
+using Random
 
+Random.seed!(10)
 nprocs() == 1 && addprocs(; exeflags = "--project")
 @everywhere include("twoblocks.jl")
 @everywhere using Main.TwoBlocksPipeline
@@ -12,5 +14,4 @@ Y = dataset[:,5] |> collect
 X = dataset[:,2:end-1]
 
 results=TwoBlocksPipeline.twoblockspipelinesearch(X,Y)
-println(results)
 println("best pipeline is: ", results[1,1],", with mean-sd ",results[1,2], " ± ",results[1,3])
