### 2022-07-03 08c1595
more info on each step of demo
