+++ src/sklearners.jl
@@ -193,7 +193,7 @@ function fit(skl::SKLearner, xx::DataFrame, y::Vector)::SKLearner
 end
 
 function transform!(skl::SKLearner, xx::DataFrame)::Vector
-	x = deepcopy(xx) |> Array
+  xx = deepcopy(xx) |> Array
   sklearner = skl.model[:sklearner]
   res = sklearner.predict(x) 
   if skl.model[:predtype] == :numeric
