+++ README.md
@@ -133,7 +133,9 @@ X = profbdata[:,2:end]
 Y = profbdata[:,1] |> Vector;
 head(x)=first(x,5)
 head(profbdata)
+```
 
+```julia
 5×7 DataFrame. Omitted printing of 1 columns
 │ Row │ Home.Away │ Favorite_Points │ Underdog_Points │ Pointspread │ Favorite_Name │ Underdog_name │
 │     │ String    │ Int64           │ Int64           │ Float64     │ String        │ String        │
@@ -194,7 +196,9 @@ Note: You can get a listing of available `Preprocessors` and `Learners` by invok
 pohe = catf |> ohe
 tr = fit_transform!(pohe,X,Y)
 head(tr)
+```
 
+```julia
 5×56 DataFrame. Omitted printing of 47 columns
 │ Row │ x1      │ x2      │ x3      │ x4      │ x5      │ x6      │ x7      │ x8      │ x9      │
 │     │ Float64 │ Float64 │ Float64 │ Float64 │ Float64 │ Float64 │ Float64 │ Float64 │ Float64 │
@@ -213,7 +217,9 @@ head(tr)
 pdec = (numf |> pca) + (numf |> ica)
 tr = fit_transform!(pdec,X,Y)
 head(tr)
+```
 
+```julia
 5×8 DataFrame
 │ Row │ x1       │ x2       │ x3       │ x4       │ x1_1       │ x2_1       │ x3_1       │ x4_1       │
 │     │ Float64  │ Float64  │ Float64  │ Float64  │ Float64    │ Float64    │ Float64    │ Float64    │
@@ -230,7 +236,9 @@ head(tr)
 ppt = (numf |> rb |> ica) + (numf |> pt |> pca)
 tr = fit_transform!(ppt,X,Y)
 head(tr)
+```
 
+```julia
 5×8 DataFrame
 │ Row │ x1          │ x2          │ x3         │ x4        │ x1_1      │ x2_1     │ x3_1       │ x4_1      │
 │     │ Float64     │ Float64     │ Float64    │ Float64   │ Float64   │ Float64  │ Float64    │ Float64   │
@@ -253,7 +261,9 @@ pred = fit_transform!(pvote,X,Y)
 sc=score(:accuracy,pred,Y)
 println(sc)
 crossvalidate(pvote,X,Y,"accuracy_score")
+```
 
+```julia
 fold: 1, 0.5373134328358209
 fold: 2, 0.7014925373134329
 fold: 3, 0.5294117647058824
@@ -298,7 +308,9 @@ prf = (numf |> rb |> pca) + (numf |> rb |> ica) + (numf |> rb |> fa) + (catf |> 
 pred = fit_transform!(prf,X,Y)
 score(:accuracy,pred,Y) |> println
 crossvalidate(prf,X,Y,"accuracy_score")
+```
 
+```julia
 fold: 1, 0.6119402985074627
 fold: 2, 0.7611940298507462
 fold: 3, 0.6764705882352942
@@ -318,7 +330,9 @@ plsvc = ((numf |> rb |> pca)+(numf |> rb |> fa)+(numf |> rb |> ica)+(catf |> ohe
 pred = fit_transform!(plsvc,X,Y)
 score(:accuracy,pred,Y) |> println
 crossvalidate(plsvc,X,Y,"accuracy_score")
+```
 
+```julia
 fold: 1, 0.6567164179104478
 fold: 2, 0.7164179104477612
 fold: 3, 0.8235294117647058
@@ -340,7 +354,9 @@ Xreg = iris[:,1:3]
 Yreg = iris[:,4] |> Vector
[…]
