+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"man/pipeline/#Pipeline","page":"Pipeline","title":"Pipeline","text":"","category":"section"},{"location":"man/pipeline/","page":"Pipeline","title":"Pipeline","text":"There are three types of Pipelines: Pipeline, ComboPipeline, and Selector Pipeline. The Pipeline (linear pipeline)  performs  sequential evaluation of fit_transform! operation to each of its elements passing the output of previous element as  input to the next element iteratively. ComboPipeline (feature union pipeline) performs dataframe concatenation of the final  outputs of its elements. Selector Pipeline acts as a selector function which outputs the results of the best learner using its internal cross-validation process.","category":"page"},{"location":"man/pipeline/","page":"Pipeline","title":"Pipeline","text":"The Pipeline uses |> symbolic expression while ComboPipeline uses +.  The expression, a |> b, is equivalent to Pipeline(a,b) function call while the expression, a + b, is equivalent to ComboPipeline(a,b). The elements a and b can be transformers, filters, learners or  pipeline themselves.","category":"page"},{"location":"man/pipeline/#Pipeline-Structure","page":"Pipeline","title":"Pipeline Structure","text":"","category":"section"},{"location":"man/pipeline/","page":"Pipeline","title":"Pipeline","text":"The linear pipeline accepts the following variables wrapped in a  Dictionary type argument:","category":"page"},{"location":"man/pipeline/","page":"Pipeline","title":"Pipeline","text":":name -> alias name for the pipeline\n:machines -> a Vector learners/transformers/pipelines\n:machine_args -> arguments to elements of the pipeline","category":"page"},{"location":"man/pipeline/","page":"Pipeline","title":"Pipeline","text":"For ease of usage, the following function calls are supported:","category":"page"},{"location":"man/pipeline/","page":"Pipeline","title":"Pipeline","text":"Pipeline(args::Dict) -> default init function\nPipeline(Vector{<:Machine},args::Dict=Dict()) -> for passing vectors of learners/transformers\nPipeline(machs::Varargs{Machine}) -> for passing learners/transformers as arguments","category":"page"},{"location":"man/pipeline/#ComboPipeline-Structure","page":"Pipeline","title":"ComboPipeline Structure","text":"","category":"section"},{"location":"man/pipeline/","page":"Pipeline","title":"Pipeline","text":"ComboPipeline or feature union pipeline accepts similar init variables with the the linear pipeline and follows similar helper functions:","category":"page"},{"location":"man/pipeline/","page":"Pipeline","title":"Pipeline","text":"ComboPipeline(args::Dict) -> default init function\nComboPipeline(Vector{<:Machine},args::Dict=Dict()) -> for passing vectors of learners/transformers\nComboPipeline(machs::Varargs{Machine}) -> for passing learners/transformers as arguments","category":"page"},{"location":"man/pipeline/#Selector-Pipeline-Structure","page":"Pipeline","title":"Selector Pipeline Structure","text":"","category":"section"},{"location":"man/pipeline/","page":"Pipeline","title":"Pipeline","text":"Selector Pipeline is based on the BestLearner ensemble. Detailed explanation can be found in:  BestLearner","category":"page"},{"location":"man/pipeline/#Macro-Functions","page":"Pipeline","title":"Macro Functions","text":"","category":"section"},{"location":"man/pipeline/","page":"Pipeline","title":"Pipeli
[…]
