+++ README.md
@@ -2,6 +2,8 @@
 
 ![Overall Stats](https://github-readme-stats.vercel.app/api?username=ppalmes&count_private=true&show_icons=true&hide=contribs)
 
+[![OpenSSF Best Practices](https://bestpractices.coreinfrastructure.org/projects/7093/badge)](https://bestpractices.coreinfrastructure.org/projects/7093)
+
 ------------------
 | **Documentation** | **Build Status** | **Help** |
 |:---:|:---:|:---:|
