+++ MAINTAINERS.md
@@ -13,4 +13,4 @@ attacks and abusive comments.
 Make sure the PR has good documentation of the why/what issues are being addressed or features added.
 
 # Vulnerability Report
-For security and vulnerability issues, please send email to Paulito Palmes.
+For security and vulnerability issues, please send email to [Paulito P. Palmes](https://github.com/ppalmes).
