### 2023-03-02 34db980
cleaning up sample/demo files
Signed-off-by: ppalmes <ppalmes@gmail.com>
