### 2023-03-02 d7a1e9b
cleaning up sample/demo files
Signed-off-by: ppalmes <ppalmes@gmail.com>
