+++ examples/juliacontalk2021.jl
@@ -1,156 +1,0 @@
-# ----------------------------------
-# Implementation
-# ----------------------------------
-
-using Distributed
-nprocs() == 1 && addprocs(; exeflags = "--project")
-workers()
-
-@everywhere using AutoMLPipeline
-@everywhere using DataFrames
-@everywhere using AutoMLPipeline: score
-@everywhere using Random
-ENV["COLUMNS"]=1000
-ENV["LINES"]=100;
-
-# disable truncation of dataframes columns
-import Base.show
-show(df::AbstractDataFrame) = show(df,truncate=0)
-show(io::IO,df::AbstractDataFrame) = show(io,df;truncate=0)
-
-# brief overview of AutoMLPipeline package
-# more details is in my Juliacon2020 presentation about this package
-# we use AutoMLPipeline to develop the pipeline optimizer
-
-# define scalers
-rb     = SKPreprocessor("RobustScaler",Dict(:name=>"rb"))
-pt     = SKPreprocessor("PowerTransformer",Dict(:name=>"pt"))
-norm   = SKPreprocessor("Normalizer",Dict(:name=>"norm"))
-mx     = SKPreprocessor("MinMaxScaler",Dict(:name=>"mx"))
-std    = SKPreprocessor("StandardScaler",Dict(:name=>"std"))
-# define extractors
-pca    = SKPreprocessor("PCA",Dict(:name=>"pca"))
-fa     = SKPreprocessor("FactorAnalysis",Dict(:name=>"fa"))
-ica    = SKPreprocessor("FastICA",Dict(:name=>"ica"))
-# define learners
-rf     = SKLearner("RandomForestClassifier",Dict(:name => "rf"))
-ada    = SKLearner("AdaBoostClassifier",Dict(:name => "ada"))
-gb     = SKLearner("GradientBoostingClassifier",Dict(:name => "gb"))
-lsvc   = SKLearner("LinearSVC",Dict(:name => "lsvc"))
-rbfsvc = SKLearner("SVC",Dict(:name => "rbfsvc"))
-dt     = SKLearner("DecisionTreeClassifier",Dict(:name =>"dt"))
-# preprocessing
-noop = Identity(Dict(:name =>"noop"))
-ohe  = OneHotEncoder(Dict(:name=>"ohe"))
-catf = CatFeatureSelector(Dict(:name=>"catf"))
-numf = NumFeatureSelector(Dict(:name=>"numf"))
-
-vscalers =    [rb,pt,norm,mx,std,noop]
-vextractors = [pca,fa,ica,noop]
-vlearners =   [rf,gb,lsvc,rbfsvc,ada,dt]
-
-dataset = getiris()
-X = dataset[:,1:4]
-Y = dataset[:,5] |> collect
-
-# one-block pipeline
-prep = @pipeline (catf |> ohe) + numf
-ppl1 = @pipeline prep |> (mx |> pca) |> rf
-crossvalidate(ppl1,X,Y)
-
-# two-block pipeline
-ppl2 = @pipeline prep |> ((norm |> ica) + (pt |> pca)) |> lsvc
-crossvalidate(ppl2,X,Y)
-
-function oneblock_pipeline_factory(scalers,extractors,learners)
-   results = @distributed (vcat) for lr in learners
-      @distributed (vcat) for xt in extractors
-         @distributed (vcat) for sc in scalers
-            # baseline preprocessing
-            prep = @pipeline ((catf |> ohe) + numf)
-            # one-block prp
-            expx = @pipeline prep |> (sc |> xt) |> lr
-            scn   = sc.name[1:end - 4];xtn = xt.name[1:end - 4]; lrn = lr.name[1:end - 4]
[…]
