+++ paper/complexity-surrogate.tex
@@ -13,12 +13,12 @@
 	 cmc          & 1    & 42.57  & 3.45 & 2.28   &  1   & 42.84  & 4.64 & 0.07   &  2   & 44.13  & 3.68 & 0.01 \\
 	 credit-app   & 1    & 11.03  & 4.59 & 0.50   &  2   & 11.60  & 4.83 & 0.03   &  2   & 12.08  & 3.80 & 0.04 \\
 	 eucalyptus   & 1    & 33.40  & 6.13 & 2.78   &  3   & 36.37  & 7.03 & 0.16   &  2   & 35.63  & 6.27 & 0.22 \\
-	 first-order  & 1    & 36.37  & 2.49 & 5.51   &  1   & 36.73  & 2.18 & 0.13   &  1   & 36.63  & 2.07 & 0.28 \\
+	 first-order  & 1*    & 36.37  & 2.49 & 5.51   &  1   & 36.73  & 2.18 & 0.13   &  1   & 36.63  & 2.07 & 0.28 \\
 	 GestureSeg   & 1    & 30.46  & 1.03 & 9.69   &  1   & 31.58  & 1.45 & 0.44   &  5   & 45.19  & 0.96 & 0.76 \\
 	 jm1          & 1    & 17.67  & 0.79 & 16.62  &  1   & 18.01  & 0.86 & 0.23   &  1   & 17.98  & 1.47 & 0.40 \\
-	 profb        & 1    & 24.69  & 5.29 & 0.61   &  1   & 24.86  & 6.02 & 0.02   &  1   & 24.41  & 4.42 & 0.04 \\
-	 plants-shape & 1    & 26.05  & 2.77 & 3.97   &  1   & 32.06  & 2.93 & 0.17   &  3   & 44.31  & 5.53 & 0.61 \\
-	 soybean      & 5    & 13.24  & 2.62 & 0.57   &  5   & 14.29  & 3.55 & 0.02   &  5   & 14.29  & 4.83 & 0.03 \\
+	 profb        & 1*    & 24.69  & 5.29 & 0.61   &  1   & 24.86  & 6.02 & 0.02   &  1   & 24.41  & 4.42 & 0.04 \\
+	 plants-shape & 1*    & 26.05  & 2.77 & 3.97   &  1   & 32.06  & 2.93 & 0.17   &  3   & 44.31  & 5.53 & 0.61 \\
+	 soybean      & 5*    & 13.24  & 2.62 & 0.57   &  5   & 14.29  & 3.55 & 0.02   &  5   & 14.29  & 4.83 & 0.03 \\
 	 \midrule                                                                                                   
 	 Median       & 1.0  & 26.05  & 2.77 & 2.28   & 1.0  & 31.58  & 3.55 & 0.07   &  2.0 & 35.63  & 3.68 & 0.04 \\
 	 \bottomrule
