+++ paper/paper.tex
@@ -359,7 +359,9 @@ results in a 4-hour time budget  discussed in the review paper
 of \cite{zoller2019}.  The comparison involves 5 AutoML approaches, namely:
 TPOT~\cite{tpot2016}, Auto-Sklearn~\cite{Feurer:AML2019Chp6,Feurer:NIPS2015},
 ATM~\cite{atm2017}, Hyperopt-Sklearn~\cite{bergstra2011nips,bergstra2012}, and
-Random Search~\cite{anderson1953}.  Among these approaches, TPOT and ATM have
+Random Search~\cite{anderson1953}.  The best results among the 5 AutoML approaches
+from each dataset  will be used as the baseline comparison in the results and discussion 
+of the succeeding section of the paper. Among these 5 approaches, TPOT and ATM have
 dominated the rankings in terms of best performance based on the
 CV classification errors on the 12 datasets as shown in the last two columns.
 
@@ -427,13 +429,27 @@ features (\texttt{numf}).
 
 \input{oneblock-all-all}
 
-Table \ref{tab:singleblock} summarizes the performance of \emph{all-all}
-strategy. The metric of performance is based on the average classification
+Table~\ref{tab:singleblock} summarizes the performance of \emph{all-all}
+strategy. The rank represents the relative performance of the proposed algo
+relative to the baseline performance of the best algorithm in Table~\ref{tab:openmldata}.
+The sign of $=$, $<$, and $>$ indicate whether the rank performance is significantly different
+from the baseline at $\alpha=0.05$ level of significance by t-test. 
+For example, rank $1, =$ indicates that the proposed algo is as good as
+the baseline in Table~\ref{tab:openmldata}. To differentiate the rank where the baseline
+or the proposed algorithm is significantly inferior or superior, the $<$ and $>$ are used, 
+respectively. For example,  among the datasets, the proposed algo in Table~\ref{tab:singleblock}
+is significantly superior to the baseline algos in \emph{first-order-theo}, \emph{profb}, and \emph{plants-shape}
+but significantly inferior in \emph{sick} and \emph{soybean} datasets. The best algos in the baseline
+has similar performance with the \emph{all-all} strategy in the rest of the datasets.
+
+\vskip 6pt
+
+The metric of performance is based on the average classification
 error (\emph{AvgErr}) using 10-fold CV. The ranking is based on its
-\emph{AvgErr} performance in comparison to the 5 AutoML approaches~\cite{zoller2019} summarized in Table~\ref{tab:openmldata}.
-The \emph{all-all} median overall rank is 1.5 using one-block pipeline.
-Its median runtime to
-perform one-block exhaustive search is 0.23 hour or 13.8 minutes per dataset.
+\emph{AvgErr} performance in comparison to the 
+5 AutoML approaches~\cite{zoller2019} summarized in Table~\ref{tab:openmldata}.
+The \emph{all-all} median overall rank is 1 using one-block pipeline.
+Its median runtime to perform one-block exhaustive search is 0.23 hour or 13.8 minutes per dataset.
 This result is encouraging because the median runtime is significantly less
 compared to the 5 AutoML algorithms with a 4-hour budget.  Closely examining 
 the runtime of each dataset, \emph{plants-shape} dataset
@@ -462,8 +478,11 @@ strategy.
 
 \input{complexity-surrogate}
 
+\vskip 6pt
+
 Tables \ref{tab:oneblock-skrf} and \ref{tab:all-one-single} show the
-performance of \emph{one-all} and \emph{all-one} strategies, respectively, for one-block MLPL. Their ranking is based on their performances against the 5 AutoML algorithms.
+performance of \emph{one-all} and \emph{all-one} strategies, 
[…]
