### 2023-05-25 #88 @ppalmes
@matbesancon @ablaom @vchuravy
I finally finished the edits based on the comments of @ablaom. Highly appreciate it if you can take a look and give advice if there are still things i missed based on the comments.

Summary of changes:
- added significance testing. Tables include rank results and a sign of either =, >, < to represent if the rank performance is similar (draw) or significantly superior or significantly inferior with alpha=0.05 level of significance.
- remove the misleading statement about MLJ

Thanks for all your inputs.

Regards,
Paulito
