### 2023-05-30 #88 @ablaom
@matbesancon

Thanks @ppalmes for considering my comments.

 I'm happy with the changes and recommend publication.
