+++ examples/AutoOfflineRL/src/offlinerls.jl
@@ -1,0 +1,219 @@
+module OfflineRLs
+
+using Distributed
+using DataFrames
+using Random
+using CSV
+using Dates
+
+using ..AbsTypes
+using ..Utils
+import ..AbsTypes: fit, fit!, transform, transform!
+
+import PythonCall
+const PYC = PythonCall
+
+using ..Utils: nested_dict_merge
+
+export DiscreteRLOffline, fit!, transform!
+export listdiscreateagents, driver
+
+const rl_dict = Dict{String, PYC.Py}()
+
+const PYRL = PYC.pynew()
+const PYPD = PYC.pynew()
+const PYNP = PYC.pynew()
+const PYDT = PYC.pynew()
+
+function __init__()
+  PYC.pycopy!(PYRL, PYC.pyimport("d3rlpy.algos"))
+  PYC.pycopy!(PYDT, PYC.pyimport("d3rlpy.dataset"))
+  PYC.pycopy!(PYPD, PYC.pyimport("pandas"))
+  PYC.pycopy!(PYNP, PYC.pyimport("numpy"))
+
+  # OfflineRLs
+  rl_dict["DiscreteBC"]           = PYRL
+  rl_dict["DQN"]                  = PYRL
+  rl_dict["NFQ"]                  = PYRL
+  rl_dict["DoubleDQN"]            = PYRL
+  rl_dict["DiscreteBCQ"]          = PYRL
+  rl_dict["DiscreteCQL"]          = PYRL
+  rl_dict["DiscreteSAC"]          = PYRL
+  #rl_dict["DiscreteRandomPolicy"] = PYRL
+end
+
+mutable struct DiscreteRLOffline <: Learner
+   name::String
+   model::Dict{Symbol,Any}
+
+   function DiscreteRLOffline(args=Dict{Symbol,Any}())
+     default_args = Dict{Symbol,Any}(
+        :name         => "DQN",
+        :tag          => "RLOffline",
+        :rlagent      => "DQN",
+        :iterations   => 100,
+        :epochs       => 10000,
+        :save_metrics => false,
+        :rlobjtrained => PYC.PyNULL,
+        :o_header     => ["day", "hour", "minute", "dow", "metric1", "metric2", "metric3", "metric4"],
+        :a_header     => ["action"],
+        :r_header     => ["reward"],
+        :save_model   => true,
+        :impl_args    => Dict{Symbol,Any}(
+            :scaler  => "min_max",
+            :use_gpu => false
+        )
+     )
+     cargs = nested_dict_merge(default_args,args)
+     datestring = Dates.format(now(), "yyyy-mm-dd-HH-MM")
+     cargs[:name] = cargs[:name]*"_"*datestring
+     rlagent = cargs[:rlagent]
+     if !(rlagent in keys(rl_dict)) 
+       println("error: $rlagent is not supported.")
+       println()
+       discreteagents()
+       error("Argument keyword error")
+     end
+     new(cargs[:name],cargs)
+   end
+end
+
+function DiscreteRLOffline(rlagent::String, args::Dict)
+  DiscreteRLOffline(Dict(:rlagent => rlagent, :name => rlagent, args...))
+end
+
+function DiscreteRLOffline(rlagent::String; args...)
+  DiscreteRLOffline(Dict(
+          :rlagent => rlagent, 
+          :name => rlagent, 
+          args...
+  ))
+end
+
+function listdiscreateagents()
+  println()
+  println("RL Discrete Agents:")
+  agents = keys(rl_dict) |> collect
+  [println("  ",agent," ") for agent in agents]
+  println("See d3rlpy python package for details about the agent arguments.")
+  nothing
+end
+
+function discreteagents()
+  println()
[…]
