### 2023-05-31 22b3357
minor edits to driver
Signed-off-by: ppalmes <ppalmes@gmail.com>
