### 2023-06-13 e5fa60f
fixes for fit function
Signed-off-by: ppalmes <ppalmes@gmail.com>
