+++ Project.toml
@@ -7,6 +7,7 @@ version = "0.4.0"
 AMLPipelineBase = "e3c3008a-8869-4d53-9f34-c96f99c8a2b6"
 CondaPkg = "992eb4ea-22a4-4c89-a5bb-47a3300528ab"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
+MicroMamba = "0b3b1443-0f03-428d-bdfb-f27f9c1191ea"
 PythonCall = "6099a3de-0909-46bc-b1f4-468b9a2dfc0d"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
