### 2023-06-13 75be4e0
fixes for fit function
Signed-off-by: ppalmes <ppalmes@gmail.com>
