+++ examples/AutoOfflineRL/examples/Project.toml
@@ -1,0 +1,6 @@
+[deps]
+AutoMLPipeline = "08437348-eef5-4817-bc1b-d4e9459680d6"
+AutoOfflineRL = "4680bba7-2b59-4a6e-a544-0ebac8b8cdd3"
+DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
+MicroMamba = "0b3b1443-0f03-428d-bdfb-f27f9c1191ea"
+Parquet = "626c502c-15b0-58ad-a749-f091afb673ae"
