### 2023-06-13 a9147d4
fixes for fit function
Signed-off-by: ppalmes <ppalmes@gmail.com>
