### 2023-06-13 bba6905
bound mamba version due to conflict with pythoncall
Signed-off-by: ppalmes <ppalmes@gmail.com>
