### 2023-06-13 993349a
fix scikitlearn in documentation deps workflow
Signed-off-by: ppalmes <ppalmes@gmail.com>
