### 2023-06-13 b1d8377
fix scikitlearn in documentation deps workflow
Signed-off-by: ppalmes <ppalmes@gmail.com>
