### 2023-06-13 5230a31
wrap pipeline search to a function
Signed-off-by: ppalmes <ppalmes@gmail.com>
