+++ examples/AutoOfflineRL/examples/script.jl
@@ -21,11 +21,11 @@ end
    numf = NumFeatureSelector();
    ## load filters
    ##### Decomposition
-   apca = SKPreprocessor("PCA",Dict(:autocomponent=>true));
+   #apca = SKPreprocessor("PCA",Dict(:autocomponent=>true,:name=>"autoPCA"));
+   #afa = SKPreprocessor("FactorAnalysis",Dict(:autocomponent=>true,:name=>"autoFA"));
+   #aica = SKPreprocessor("FastICA",Dict(:autocomponent=>true,:name=>"autoICA"));
    pca = SKPreprocessor("PCA");
-   afa = SKPreprocessor("FactorAnalysis",Dict(:autocomponent=>true));
    fa = SKPreprocessor("FactorAnalysis");
-   aica = SKPreprocessor("FastICA",Dict(:autocomponent=>true));
    ica = SKPreprocessor("FastICA");
    noop = Identity(Dict(:name => "Noop"));
 end
@@ -40,36 +40,38 @@ reward = df[:,["reward"]] |> deepcopy |> DataFrame
 action = df[:,["action"]] |> deepcopy |> DataFrame
 action_reward = DataFrame[action, reward]
 
-agentnames = ["DiscreteCQL","NFQ","DoubleDQN","DiscreteSAC","DiscreteBCQ","DiscreteBC","DQN"]
-scalers =  [rb,pt,norm,std,mx,noop]
-extractors = [pca,ica,fa,noop];
-dfresults = @sync @distributed (vcat) for agentname in agentnames
-   @distributed (vcat) for sc in scalers
-      @distributed (vcat) for xt  in extractors
-         try
-            rlagent = DiscreteRLOffline(agentname,Dict(:runtime_args=>Dict(:n_epochs=>1)))
-            rlpipeline = ((numf |> sc |> xt)) |> rlagent 
-            res = fit_transform!(rlpipeline,df_input,action_reward)
-            s = [x.value[1] for x in res] |> sum
-            scn   = sc.name[1:end - 4]; xtn = xt.name[1:end - 4]; lrn = rlagent.name[1:end - 4]
-            pname = "$scn |> $xtn |> $lrn"
-            if !isnan(s)
-               DataFrame(pipeline=pname,perf=s)
-            else
+function pipelinesearch()
+   agentnames = ["DiscreteCQL","NFQ","DoubleDQN","DiscreteSAC","DiscreteBCQ","DiscreteBC","DQN"]
+   scalers =  [rb,pt,norm,std,mx,noop]
+   extractors = [pca,ica,fa,noop]
+   dfresults = @sync @distributed (vcat) for agentname in agentnames
+      @distributed (vcat) for sc in scalers
+         @distributed (vcat) for xt  in extractors
+            try
+               rlagent = DiscreteRLOffline(agentname,Dict(:runtime_args=>Dict(:n_epochs=>1)))
+               rlpipeline = ((numf |> sc |> xt)) |> rlagent 
+               res = fit_transform!(rlpipeline,df_input,action_reward)
+               s = [x.value[1] for x in res] |> sum
+               scn   = sc.name[1:end - 4]; xtn = xt.name[1:end - 4]; lrn = rlagent.name[1:end - 4]
+               pname = "$scn |> $xtn |> $lrn"
+               if !isnan(s)
+                  DataFrame(pipeline=pname,perf=s)
+               else
+                  DataFrame()
+               end
+            catch e
+               println("error in $agentname")
+               #DataFrame(agent=agentname,scaler=sc,xtractor=xt,perf=NaN)
                DataFrame()
             end
-         catch e
-            println("error in $agentname")
[…]
