+++ Project.toml
@@ -16,7 +16,7 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 AMLPipelineBase = "0.1"
 CondaPkg = "0.2"
 DataFrames = "0.17, 0.18, 0.19, 0.20, 0.21, 0.22, 1.0, 1.1"
-MicroMamba = "0.1.0 - 0.1.12"
+MicroMamba = "0.1.0 - 0.1.12, 0.1"
 PythonCall = "0.9"
 julia = "1"
 
