### 2023-06-14 #129 @whedon
[REVIEW]: Designing Machine Learning Pipeline Toolkit for AutoML Surrogate Modeling Optimization
**Submitting author:** @ppalmes (<a href="http://orcid.org/0000-0002-3145-6356">Paulito Palmes</a>)
**Repository:** <a href="https://github.com/IBM/AutoMLPipeline.jl" target ="_blank">https://github.com/IBM/AutoMLPipeline.jl</a>
**Version:** v0.4.1
