### 2023-06-14 #129 @matbesancon
@ppalmes in the meantime can you see if you can either compress some parts to get under 6 pages, and if not, maybe expanding a bit to make the 7th page worth having :)
