### 2023-06-14 #129 @matbesancon
@ppalmes in other thing to do: archiving the repository on zenodo and setting the corresponding DOI
