### 2023-06-14 50f57bf
paper edits for doi entries
Signed-off-by: ppalmes <ppalmes@gmail.com>
