+++ paper/paper.tex
@@ -30,7 +30,8 @@
 \def\lstfloatautorefname{Listing}
 \usepackage{hyperref}
 \newcommand{\theHalgorithm}{\arabic{algorithm}}
-
+\usepackage{doi}
+\usepackage{hyperref}
  
 \begin{document}
 
@@ -476,6 +477,7 @@ strategy.
 
 \input{twoblocks-all-one}
 
+
 \input{complexity-surrogate}
 
 \vskip 6pt
@@ -670,7 +672,7 @@ sharing of applications to wider fields of study.
 
 It is with this background that AMLP was developed. It started as a hobby to
 learn \emph{Julia} by implementing ML algorithms.
-Due to it's high performance, \emph{Julia} provides a great
+Due to its high performance, \emph{Julia} provides a great
 playground to test ideas and gain results easily,
 which then triggers the desire to test more ideas, all achievable in a short amount of time.
 The resulting explorations can easily be turned into a package which is highly
