+++ paper/oneblock-all-one.tex
@@ -1,7 +1,7 @@
 \begin{table*}[htbp]
   \centering
   \tbl{\emph{All-one} (one-block)}{
-  \tiny
+  \small
   \begin{tabular}{{@{} rcrrrlll @{}}}
 	 \toprule
 	 & & & & & \multicolumn{2}{c}{Block} \\
