+++ paper/oneblock-all-all.tex
@@ -1,7 +1,7 @@
 \begin{table*}[htbp]
   \tbl{\emph{All-all} (one-block) search strategy}{
   \centering
-  \tiny
+  \small
   \begin{tabular}{{@{} rcrrrlll @{}}}
     \toprule
 	 & & & & & \multicolumn{2}{c}{Block} \\
