+++ paper/twoblocks-all-one.tex
@@ -1,7 +1,7 @@
 \begin{table*}[htbp]
   \tbl{\emph{All-one} (two-block)}{
   \centering
-  \tiny
+  \small
   \begin{tabular}{{@{} rcrrr|ll|ll|l @{}}}
 	 \toprule
 	 & & & & & \multicolumn{2}{|c|}{Block 1} & \multicolumn{2}{c|}{Block 2} \\
