### 2023-06-14 cbf4589
paper edits for doi entries
Signed-off-by: ppalmes <ppalmes@gmail.com>
