+++ paper/oneblock-skrf.tex
@@ -1,7 +1,7 @@
 \begin{table*}[htbp]
   \centering
   \tbl{\emph{One-all} (one-block)}{
-  \tiny
+  \small
   \begin{tabular}{{@{} rcrrrlll @{}}}
 	 \toprule
 	 & & & & & \multicolumn{2}{c}{Block} \\
