+++ paper/complexity-surrogate.tex
@@ -1,7 +1,7 @@
 \begin{table*}[htbp]
   \tbl{\emph{All-one with Surrogates}}{
   \centering
-  \tiny
+  \small
   \begin{tabular}{{@{} r|crrr|crrr|crrr| @{}}}
 	 \toprule
 	              & \multicolumn{4}{c|}{Baseline: All-one} & \multicolumn{4}{c|}{Surrogate: Baseline+PRP} & \multicolumn{4}{c|}{Surrogates: Baseline+PRP+LR} \\
