+++ paper/paper.tex
@@ -94,15 +94,15 @@ features:
       \emph{scikit-learn}~\cite{pedregosa2011},
       \emph{caret}~\cite{caret2008}, etc.
 \item Easily extensible architecture by
-      overloading two interfaces: \texttt{fit!} and \texttt{transform!}
+      overloading two APIs: \texttt{fit!} and \texttt{transform!}
 \item High-level implementation of meta-ensembles
-      for the ensemble composition of ensembles (recursively if needed) for robust prediction routines.
+      for the composition of ensembles (recursively if needed) to support robust prediction routines.
 \item Categorical and numerical feature selectors
       for specialized preprocessing routines based on types.
 \item High-level code of parallelized cross-validation (CV)
       routines by leveraging on the multi-threading and
       distributed computing features
-      in \emph{Julia}~\cite{bezanson2017julia}.
+      of the \emph{Julia} language~\cite{bezanson2017julia}.
 \end{itemize}
 
 AMLP's main feature is the use of relatively compact symbolic expression in
@@ -110,10 +110,10 @@ pipeline composition.  For instance, a pipeline expression to extract the
 numerical features (\texttt{numf}) for 
 \texttt{pca} decomposition, concatenated with one-hot
 encoding (\texttt{ohe}) of categorical features (\texttt{catf})
-of a given data for RandomForest (\texttt{rf}) modeling can be written
+of a given data for RandomForest (\texttt{rf}) modeling can be expressed
 as:
 \begin{lstlisting}[language=Julia,numbers=none]
-   pipe = (numf |> pca) + (catf |> ohe) |> rf 
+   (numf |> pca) + (catf |> ohe) |> rf 
 \end{lstlisting}
 
 \vskip 6pt
@@ -137,8 +137,8 @@ AMLP's main design was inspired by the
 having common APIs across different ML implementations, \emph{Orchestra} showed
 that one can mix and match ML preprocessing routines from \emph{caret},
 \emph{scikit-learn}, and \emph{Julia} seamlessly in a pipeline by leveraging on
-\emph{PyCall}~\cite{pycall} and \emph{RCall}~\cite{rcall} wrappers for
-\emph{Python} and \emph{R}, respectively. This is a convenient package
+\emph{PyCall}~\cite{pycall} and \emph{RCall}~\cite{rcall} wrapper libraries for
+\emph{Python} and \emph{R}, respectively. \emph{Orchestra}  is a convenient package
 because there are many cases where implementations of ML functions can be found
 exclusively in either \emph{scikit-learn} or \emph{caret}.  Having both
 libraries available for perusal is a great productivity boost for machine
@@ -159,9 +159,9 @@ interoperability.  \emph{Lale} has consistent high-level interface for popular
 AutoML algorithms such as:
 \emph{Hyperopt}~\cite{bergstra2011nips,bergstra2012},
 \emph{GridSearchCV}~\cite{pedregosa2011}, and
-\emph{SMAC}~\cite{hutter2011smac}.  It makes it possible to have an
-easy comparison of the different AutoML algorithms for benchmarking, and research,
-as well as for applications. While AMLP is at the early stage of development, its
+\emph{SMAC}~\cite{hutter2011smac}.  \emph{Lale} supports
+easy comparison of the different AutoML algorithms for benchmarking and research.
+While AMLP is at the early stage of development, its
 pipeline architecture can be used as building blocks for the development of
 future AutoML algorithms similar to \emph{Lale} in \emph{Julia}'s
 ecosystem.
@@ -169,9 +169,9 @@ ecosystem.
 \section{AMLP Architecture}
 \label{amlparchitecture}
 
-The code in Listing \ref{lst:abstypes} describes the abstract type hierarchy
[…]
