+++ paper/amlp.bib
@@ -155,6 +155,16 @@
   Title = {Pro Football Scores},
   Year = {2014}}
 
+  @inproceedings{avatar2020,
+     Author = {Tien-Dung Nguyen and Tomasz Maszczyk and  Katarzyna Musial and Marc-Andre Z{\"o}ller and Bogdan Gabrys},
+     Booktitle = {Advances in Intelligent Data Analysis XVIII},
+     Date-Added = {2020-05-29 08:38:02 +0100},
+     Date-Modified = {2020-05-29 08:50:59 +0100},
+     Pages = {352--365},
+     Publisher = {Springer International Publishing},
+     Title = {AVATAR - Machine Learning Pipeline Evaluation Using Surrogate Model},
+     Doi = {10.1007/978-3-030-44584-3\_28},
+     Year = {2020}}
 
 @article{openml,
   Author = {J. Vanschoren and J.N. van Rijn and B. Bischl and L. Torgo},
