+++ paper/paper.tex
@@ -30,7 +30,6 @@
 \def\lstfloatautorefname{Listing}
 \usepackage{hyperref}
 \newcommand{\theHalgorithm}{\arabic{algorithm}}
-\usepackage{doi}
 \usepackage{hyperref}
  
 \begin{document}
