+++ examples/AutoOfflineRL/Project.toml
@@ -14,6 +14,7 @@ MicroMamba = "0b3b1443-0f03-428d-bdfb-f27f9c1191ea"
 Parquet = "626c502c-15b0-58ad-a749-f091afb673ae"
 PythonCall = "6099a3de-0909-46bc-b1f4-468b9a2dfc0d"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 
 [compat]
 julia = "1"
