+++ AutoOfflineRL/src/offlinerls.jl
@@ -1,0 +1,279 @@
+module OfflineRLs
+
+using AutoOfflineRL
+using PythonCall
+import Statistics
+using Parquet
+using Distributed
+using DataFrames: DataFrame, dropmissing
+using Random
+using CSV
+using Dates
+
+using ..AbsTypes
+using ..Utils
+import ..AbsTypes: fit, fit!, transform, transform!
+
+import PythonCall
+const PYC = PythonCall
+
+using ..Utils: nested_dict_merge
+
+export DiscreteRLOffline, fit!, transform!, fit, transform
+export listdiscreateagents, driver
+export crossvalidateRL
+
+const rl_dict = Dict{String, PYC.Py}()
+const metric_dict = Dict{String, PYC.Py}()
+
+const PYRL = PYC.pynew()
+const PYPD = PYC.pynew()
+const PYNP = PYC.pynew()
+const PYDT = PYC.pynew()
+const PYMT = PYC.pynew()
+const PYSK  = PYC.pynew()
+
+
+function __init__()
+  PYC.pycopy!(PYRL, PYC.pyimport("d3rlpy.algos"))
+  PYC.pycopy!(PYDT, PYC.pyimport("d3rlpy.datasets"))
+  PYC.pycopy!(PYSK, PYC.pyimport("sklearn.model_selection"))
+  PYC.pycopy!(PYMT, PYC.pyimport("d3rlpy.metrics"))
+  PYC.pycopy!(PYPD, PYC.pyimport("pandas"))
+  PYC.pycopy!(PYNP, PYC.pyimport("numpy"))
+
+  # OfflineRLs
+  metric_dict["cross_validate"]   = PYSK.cross_validate
+  metric_dict["train_test_split"] = PYSK.train_test_split
+  metric_dict["td_error_scorer"]  = PYMT.td_error_scorer
+  metric_dict["average_value_estimation_scorer"]  = PYMT.average_value_estimation_scorer
+  metric_dict["get_cartpole"] = PYDT.get_cartpole
+
+  rl_dict["DiscreteBC"]           = PYRL
+  rl_dict["DQN"]                  = PYRL
+  rl_dict["NFQ"]                  = PYRL
+  rl_dict["DoubleDQN"]            = PYRL
+  rl_dict["DiscreteBCQ"]          = PYRL
+  rl_dict["DiscreteCQL"]          = PYRL
+  rl_dict["DiscreteSAC"]          = PYRL
+  #rl_dict["DiscreteRandomPolicy"] = PYRL
+end
+
+mutable struct DiscreteRLOffline <: Learner
+   name::String
+   model::Dict{Symbol,Any}
+
+   function DiscreteRLOffline(args=Dict{Symbol,Any}())
+     default_args = Dict{Symbol,Any}(
+        :name         => "DQN",
+        :tag          => "RLOffline",
+        :rlagent      => "DQN",
+        :iterations   => 100,
+        :save_metrics => false,
+        :rlobjtrained => PYC.PyNULL,
+        :o_header     => ["day", "hour", "minute", "dow", "metric1", "metric2", "metric3", "metric4"],
+        :a_header     => ["action"],
+        :r_header     => ["reward"],
+        :save_model   => false,
+        :runtime_args => Dict{Symbol, Any}(
+            :n_epochs => 5,
+        ),
+        :impl_args    => Dict{Symbol,Any}(
+            :scaler  => "min_max",
+            :use_gpu => false,
+        )
+     )
+     cargs = nested_dict_merge(default_args,args)
+     #datestring = Dates.format(now(), "yyyy-mm-dd-HH-MM")
+     cargs[:name] = cargs[:name]*"_"*randstring(3)
+     rlagent = cargs[:rlagent]
[…]
