+++ examples/AutoOfflineRL/examples/test.py
@@ -1,30 +1,0 @@
-from d3rlpy.datasets import get_cartpole
-from d3rlpy.algos import DQN
-from d3rlpy.metrics.scorer import td_error_scorer
-from d3rlpy.metrics.scorer import average_value_estimation_scorer
-from d3rlpy.metrics.scorer import evaluate_on_environment
-from sklearn.model_selection import train_test_split
-
-dataset, env = get_cartpole()
-train_episodes, test_episodes = train_test_split(dataset)
-dqn = DQN()
-dqn.fit(train_episodes,
-        eval_episodes=test_episodes,
-        n_epochs=1,
-        scorers={
-            'td_error': td_error_scorer,
-            'value_scale': average_value_estimation_scorer
-        }
-       )
-
-
-from sklearn.model_selection import cross_validate
-
-scores = cross_validate(dqn,
-                        dataset,
-                        scoring={
-                            'td_error': td_error_scorer,
-                            'value_scale': average_value_estimation_scorer
-                        },
-                        fit_params={'n_epochs':1}
-                       )
