+++ examples/AutoOfflineRL/CondaPkg.toml
@@ -1,10 +1,0 @@
-channels = ["mkl", "scikit-learn"]
-
-[deps]
-scikit-learn = ""
-pandas = ""
-numpy = ""
-python = ""
-
-[pip.deps]
-d3rlpy = ""
