### 2023-06-21 15787cd
crossvalidation data and algo
Signed-off-by: ppalmes <ppalmes@gmail.com>
