+++ AutoOfflineRL/src/offlinerls.jl
@@ -46,6 +46,7 @@ function __init__()
   metric_dict["cross_validate"]   = PYSK.cross_validate
   metric_dict["train_test_split"] = PYSK.train_test_split
   metric_dict["td_error_scorer"]  = PYMT.td_error_scorer
+  metric_dict["discrete_action_match_scorer"] = PYMT.discrete_action_match_scorer
   metric_dict["average_value_estimation_scorer"]  = PYMT.average_value_estimation_scorer
   metric_dict["get_cartpole"] = PYDT.get_cartpole
 
@@ -76,7 +77,7 @@ mutable struct DiscreteRLOffline <: Learner
         :r_header     => ["reward"],
         :save_model   => false,
         :runtime_args => Dict{Symbol, Any}(
-            :n_epochs => 5,
+            :n_epochs => 3,
         ),
         :impl_args    => Dict{Symbol,Any}(
             :scaler  => "min_max",
@@ -126,15 +127,11 @@ function discreteagents()
   println("use: listdiscreateagents() to get the available RL agents")
 end
 
-function createmdpdata!(agent::DiscreteRLOffline, df::DataFrame, action_reward::Vector)
+function createmdpdata!(agent::DiscreteRLOffline, df::DataFrame, action_reward_term::Vector)
   _observations = df |> Array .|> PYC.float |> x -> PYNP.array(x, dtype = "float32")
-  _actions      = action_reward[1] |> Array .|> PYC.float |> x -> PYNP.array(x, dtype = "float32")
-  _rewards      = action_reward[2] |> Array .|> PYC.float |> x -> PYNP.array(x, dtype = "float32")
-  ## inject end of data by terminal column
-  nrow, _         = size(df)
-  _terminals      = zeros(Int, nrow)
-  _terminals[end] = 1
-  _terminals      = _terminals .|> PYC.float |> x -> PYNP.array(x, dtype = "int32")
+  _actions      = action_reward_term[1] |> Array .|> PYC.float |> x -> PYNP.array(x, dtype = "float32")
+  _rewards      = action_reward_term[2] |> Array .|> PYC.float |> x -> PYNP.array(x, dtype = "float32")
+  _terminals      = action_reward_term[3] |> Array .|> PYC.float |> x -> PYNP.array(x, dtype = "float32")
   ## create dataset for RLOffline
   mdp_dataset = PYDT.MDPDataset(
     observations = _observations,
@@ -159,12 +156,12 @@ function checkheaders(agent::DiscreteRLOffline, df)
    for header in vcat(o_header, a_header, r_header)]
 end
 
-function fit!(agent::DiscreteRLOffline, df::DataFrame, action_reward::Vector)::Nothing
+function fit!(agent::DiscreteRLOffline, df::DataFrame, action_reward_term::Vector)::Nothing
   # check if headers exist
   #checkheaders(agent::DiscreteRLOffline, df)
   # create mdp data
   nrow, ncol  = size(df)
-  mdp_dataset = createmdpdata!(agent, df,action_reward)
+  mdp_dataset = createmdpdata!(agent, df,action_reward_term)
   ## prepare algorithm
   runtime_args = agent.model[:runtime_args]
   logging    = agent.model[:save_metrics]
@@ -206,7 +203,7 @@ function transform!(agent::DiscreteRLOffline,df::DataFrame=DataFrame())::Vector
   return res
 end
 
-function prp_fit_transform(pipe::Machine, instances::DataFrame,actreward::Vector)
+function prp_fit_transform(pipe::Machine, instances::DataFrame,actrewterm::Vector)
    machines = pipe.model[:machines]
    machine_args = pipe.model[:machine_args]
 
[…]
