### 2023-06-21 98f5ca2
crossvalidation data and algo
Signed-off-by: ppalmes <ppalmes@gmail.com>
